# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Order Fulfillment and Logistics Hub API
  version: 0.1.0
  description: >-
    Contract for BUSINESS administration, trusted internal intake, and isolated
    DemoScenario operations. JSON field names use snake_case; identifiers and
    NUMERIC(18,3) quantities are strings.
servers:
  - url: http://localhost:8080
    description: Local Docker Compose
...
security:
  - basicAuth: []
paths:
...
  /api/v1/fulfillments:
    get:
      tags: [Fulfillments]
      operationId: listFulfillments
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
        - $ref: '#/components/parameters/DateFrom'
        - $ref: '#/components/parameters/DateTo'
        - name: provider_id
          in: query
          schema: { $ref: '#/components/schemas/Identifier' }
        - name: shipping_progress
          in: query
          schema: { $ref: '#/components/schemas/ShippingProgress' }
        - name: outcome
          in: query
          schema: { $ref: '#/components/schemas/FulfillmentOutcome' }
      responses:
        '200':
          description: Fulfillment task page
          content:
            application/json:
              schema: { $ref: '#/components/schemas/FulfillmentPage' }
        default: { $ref: '#/components/responses/ErrorResponse' }
...
  /api/v1/products:
    get:
      tags: [MasterData]
      operationId: listProducts
      parameters: [{ $ref: '#/components/parameters/Page' }, { $ref: '#/components/parameters/Size' }]
      responses:
        '200':
          description: Products
          content: { application/json: { schema: { $ref: '#/components/schemas/MasterDataPage' } } }
        default: { $ref: '#/components/responses/ErrorResponse' }
    post:
      tags: [MasterData]
      operationId: createProduct
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/Operator'
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/ProductWrite' } } }
      responses:
        '201':
          description: Product created
          content: { application/json: { schema: { $ref: '#/components/schemas/MasterDataRecord' } } }
        default: { $ref: '#/components/responses/ErrorResponse' }
...
  /api/v1/connectors:
    get:
      tags: [Connectors]
      operationId: listConnectors
      responses:
        '200':
          description: Four source-channel connector configurations
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/ConnectorConfig' }
        default: { $ref: '#/components/responses/ErrorResponse' }
  /api/v1/connectors/{source_channel}:
    parameters:
      - $ref: '#/components/parameters/SourceChannelPath'
    get:
      tags: [Connectors]
      operationId: getConnector
      responses:
        '200':
          description: Connector mode and non-secret configuration
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ConnectorConfig' }
        default: { $ref: '#/components/responses/ErrorResponse' }
    patch:
      tags: [Connectors]
      operationId: updateConnector
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/Operator'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ConnectorPatch' }
      responses:
        '200':
          description: Non-secret connector configuration updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ConnectorConfig' }
        default: { $ref: '#/components/responses/ErrorResponse' }
  /api/v1/connectors/{source_channel}/test-connection:
    parameters:
      - $ref: '#/components/parameters/SourceChannelPath'
    post:
      tags: [Connectors]
      operationId: testConnectorConnection
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/Operator'
      responses:
        '200':
          description: Mock or real adapter connection result; does not advance orders
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ConnectionTestResult' }
        default: { $ref: '#/components/responses/ErrorResponse' }
...
  /api/v1/analytics/channels:
    get:
      tags: [Analytics]
      operationId: getChannelAnalytics
      parameters:
        - $ref: '#/components/parameters/DateFrom'
        - $ref: '#/components/parameters/DateTo'
        - name: source_channel
          in: query
          schema: { $ref: '#/components/schemas/SourceChannel' }
      responses:
        '200':
          description: Channel by natural day metrics
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/ChannelMetric' }
        default: { $ref: '#/components/responses/ErrorResponse' }
  /api/v1/analytics/products:
    get:
      tags: [Analytics]
      operationId: getProductAnalytics
      parameters:
        - $ref: '#/components/parameters/DateFrom'
        - $ref: '#/components/parameters/DateTo'
        - name: source_channel
          in: query
          schema: { $ref: '#/components/schemas/SourceChannel' }
        - name: product_id
          in: query
          schema: { $ref: '#/components/schemas/Identifier' }
        - name: sku_id
          in: query
          schema: { $ref: '#/components/schemas/Identifier' }
        - name: category_id
          in: query
          schema: { $ref: '#/components/schemas/Identifier' }
      responses:
        '200':
          description: Canonical SKU units; bundles expanded to components
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/ProductMetric' }
        default: { $ref: '#/components/responses/ErrorResponse' }
  /api/v1/analytics/fulfillments:
    get:
      tags: [Analytics]
      operationId: getFulfillmentAnalytics
      parameters:
        - $ref: '#/components/parameters/DateFrom'
        - $ref: '#/components/parameters/DateTo'
        - name: source_channel
          in: query
          description: Limits fulfillment, backlog, funnel, and tracking metrics to one source channel.
          schema: { $ref: '#/components/schemas/SourceChannel' }
        - name: provider_id
          in: query
          schema: { $ref: '#/components/schemas/Identifier' }
      responses:
        '200':
          description: Fulfillment, stock-out, procurement, waiting, and sync metrics
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/FulfillmentMetric' }
        default: { $ref: '#/components/responses/ErrorResponse' }
...
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
      description: Nginx authenticates management, internal, and demo operations; explicitly public connector callbacks use their own signatures.
  headers:
    RequestId:
      description: Request correlation identifier
      schema: { type: string }
  responses:
    ErrorResponse:
      description: Unified business or technical error
      headers:
        X-Request-Id: { $ref: '#/components/headers/RequestId' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
  parameters:
    Page:
      name: page
      in: query
      schema: { type: integer, minimum: 0, default: 0 }
    Size:
      name: size
      in: query
      schema: { type: integer, minimum: 1, maximum: 200, default: 20 }
...
    DateFrom:
      name: date_from
      in: query
      schema: { type: string, format: date }
    DateTo:
      name: date_to
      in: query
      schema: { type: string, format: date }
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: { type: string, minLength: 8, maxLength: 255 }
    Operator:
      name: X-Operator
      in: header
      required: true
      description: Required at the backend seam; the public gateway overwrites this value with the authenticated principal.
      schema: { type: string, minLength: 1, maxLength: 128 }
...
    SourceChannelPath:
      name: source_channel
      in: path
      required: true
      schema: { $ref: '#/components/schemas/SourceChannel' }
  schemas:
    Identifier:
      type: string
      pattern: '^[1-9][0-9]*$'
      example: '42'
    DecimalQuantity:
      type: string
      pattern: '^(0|[1-9][0-9]*)(\.[0-9]{1,3})?$'
      example: '6.000'
...
    SourceChannel:
      type: string
      enum: [CAISHIXIAN, JUFUBAO, FEIXIANG, WECOM]
...
    ShippingProgress:
      type: string
      enum: [NOT_SHIPPED, PARTIALLY_SHIPPED, SHIPPED]
    FulfillmentOutcome:
      type: string
      enum: [IN_PROGRESS, FULLY_FULFILLED, PARTIALLY_FULFILLED, CANCELLED]
...
    PageMeta:
      type: object
      required: [page, size, total_elements, total_pages]
      properties:
        page: { type: integer, minimum: 0 }
        size: { type: integer, minimum: 1 }
        total_elements: { type: integer, format: int64, minimum: 0 }
        total_pages: { type: integer, minimum: 0 }
    FieldError:
      type: object
      required: [field, code, message]
      properties:
        field: { type: string }
        code: { type: string }
        message: { type: string }
    ApiError:
      type: object
      required: [business_code, message, http_status, request_id, trace_id]
      properties:
        business_code: { type: string, maxLength: 64 }
        message: { type: string }
        http_status: { type: integer, minimum: 400, maximum: 599 }
        request_id: { type: string }
        trace_id: { type: string }
        field_errors:
          type: array
          items: { $ref: '#/components/schemas/FieldError' }
        details:
          type: object
          additionalProperties: true
...
    Fulfillment:
      type: object
      required: [id, fulfillment_no, order_line_id, provider_id, requested_quantity, cumulative_shipped_quantity, cancelled_quantity, shipping_progress, outcome, version]
      properties:
        id: { $ref: '#/components/schemas/Identifier' }
        fulfillment_no: { type: string }
        order_line_id: { $ref: '#/components/schemas/Identifier' }
        provider_id: { $ref: '#/components/schemas/Identifier' }
        requested_quantity: { $ref: '#/components/schemas/DecimalQuantity' }
        cumulative_shipped_quantity: { $ref: '#/components/schemas/DecimalQuantity' }
        cancelled_quantity: { $ref: '#/components/schemas/DecimalQuantity' }
        shipping_progress: { $ref: '#/components/schemas/ShippingProgress' }
        outcome: { $ref: '#/components/schemas/FulfillmentOutcome' }
        exception_code: { type: string }
        exception_reason: { type: string }
        version: { type: integer, format: int64, minimum: 0 }
...
    FulfillmentPage:
      allOf:
        - $ref: '#/components/schemas/PageMeta'
        - type: object
          required: [items]
          properties:
            items:
              type: array
              items: { $ref: '#/components/schemas/Fulfillment' }
...
    MasterDataRecord:
      type: object
      required: [id, code, name, active, version]
      properties:
        id: { $ref: '#/components/schemas/Identifier' }
        code: { type: string }
        name: { type: string }
        active: { type: boolean }
        version: { type: integer, minimum: 0 }
        attributes: { type: object, additionalProperties: true }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    MasterDataPage:
      allOf:
        - $ref: '#/components/schemas/PageMeta'
        - type: object
          required: [items]
          properties:
            items:
              type: array
              items: { $ref: '#/components/schemas/MasterDataRecord' }
...
    ProductWrite:
      type: object
      required: [product_code, product_name, category_id]
      properties:
        product_code: { type: string, minLength: 1, maxLength: 64 }
        product_name: { type: string, minLength: 1, maxLength: 128 }
        category_id: { $ref: '#/components/schemas/Identifier' }
        active: { type: boolean, default: true }
...
    ConnectorConfig:
      type: object
      required: [source_channel, client_mode, transport_mode, enabled, version]
      properties:
        source_channel: { $ref: '#/components/schemas/SourceChannel' }
        client_mode: { type: string, enum: [MOCK, REAL], description: Maps to connector_configs.mode. }
        transport_mode: { type: string, enum: [EXCEL, API] }
        enabled: { type: boolean }
        endpoint: { type: string, format: uri, nullable: true }
        credential_configured: { type: boolean }
        version: { type: integer, minimum: 0 }
    ConnectorPatch:
      type: object
      required: [expected_version]
      properties:
        expected_version: { type: integer, minimum: 0 }
        client_mode: { type: string, enum: [MOCK, REAL] }
        transport_mode: { type: string, enum: [EXCEL, API] }
        enabled: { type: boolean }
        endpoint: { type: string, format: uri }
        credential_secret_ref: { type: string, description: Secret manager reference; never return its value. }
      minProperties: 2
    ConnectionTestResult:
      type: object
      required: [success, checked_at]
      properties:
        success: { type: boolean }
        checked_at: { type: string, format: date-time }
        latency_ms: { type: integer, minimum: 0 }
        business_code: { type: string, nullable: true }
        message: { type: string, nullable: true }
...
    ChannelMetric:
      type: object
      required: [metric_date, source_channel, order_count, order_line_count, actual_shipped_quantity, canonical_quantity, shipped_quantity, shipment_count, exception_order_count, out_of_stock_order_count, sync_failed_count]
      properties:
        metric_date: { type: string, format: date }
        source_channel: { $ref: '#/components/schemas/SourceChannel' }
        order_count: { type: integer, minimum: 0 }
        order_line_count: { type: integer, minimum: 0 }
        actual_shipped_quantity: { $ref: '#/components/schemas/DecimalQuantity' }
        canonical_quantity: { $ref: '#/components/schemas/DecimalQuantity' }
        shipped_quantity: { $ref: '#/components/schemas/DecimalQuantity' }
        shipment_count: { type: integer, minimum: 0 }
        exception_order_count: { type: integer, minimum: 0 }
        out_of_stock_order_count: { type: integer, minimum: 0 }
        sync_failed_count: { type: integer, minimum: 0 }
    ProductMetric:
      type: object
      required: [metric_date, source_channel, product_id, product_code, product_name, sku_id, sku_code, sku_name, order_count, shipment_count, actual_shipped_quantity, canonical_quantity, shipped_quantity]
      properties:
        metric_date: { type: string, format: date }
        source_channel: { $ref: '#/components/schemas/SourceChannel' }
        category_id: { $ref: '#/components/schemas/Identifier' }
        category_code: { type: string }
        category_name: { type: string }
        product_id: { $ref: '#/components/schemas/Identifier' }
        product_code: { type: string }
        product_name: { type: string }
        sku_id: { $ref: '#/components/schemas/Identifier' }
        sku_code: { type: string }
        sku_name: { type: string }
        order_count: { type: integer, minimum: 0 }
        shipment_count: { type: integer, minimum: 0 }
        actual_shipped_quantity: { $ref: '#/components/schemas/DecimalQuantity' }
        canonical_quantity: { $ref: '#/components/schemas/DecimalQuantity' }
        shipped_quantity: { $ref: '#/components/schemas/DecimalQuantity' }
        source_mappings:
          type: array
          description: Active source SKU and package multipliers for this channel and Canonical SKU.
          items:
            type: object
            required: [source_sku_ref]
            properties:
              source_sku_ref: { type: string }
              source_product_name: { type: string }
              source_specification: { type: string }
              quantity_multiplier: { $ref: '#/components/schemas/DecimalQuantity' }
        jd_sku_codes:
          type: array
          items: { type: string }
    FulfillmentMetric:
      type: object
      required: [metric_date, source_channel, provider_id, provider_code, provider_name, provider_type, fulfillment_count, fulfilled_quantity, not_shipped_count, partially_shipped_count, fully_shipped_count, procurement_ticket_count, out_of_stock_fulfillment_count, awaiting_shipment_count, shipped_shipment_count, awaiting_tracking_count, tracking_received_count, awaiting_sync_count, sync_failed_count, synced_count, shipment_count, shipped_quantity, average_tracking_hours]
      properties:
        metric_date: { type: string, format: date }
        source_channel: { $ref: '#/components/schemas/SourceChannel' }
        provider_id: { $ref: '#/components/schemas/Identifier' }
        provider_code: { type: string }
        provider_name: { type: string }
        provider_type: { type: string, enum: [JD_WAREHOUSE, THIRD_PARTY] }
        fulfillment_count: { type: integer, minimum: 0 }
        fulfilled_quantity: { $ref: '#/components/schemas/DecimalQuantity' }
        not_shipped_count: { type: integer, minimum: 0 }
        partially_shipped_count: { type: integer, minimum: 0 }
        fully_shipped_count: { type: integer, minimum: 0 }
        procurement_ticket_count: { type: integer, minimum: 0 }
        out_of_stock_fulfillment_count: { type: integer, minimum: 0 }
        awaiting_shipment_count: { type: integer, minimum: 0 }
        shipped_shipment_count: { type: integer, minimum: 0 }
        awaiting_tracking_count: { type: integer, minimum: 0 }
        tracking_received_count: { type: integer, minimum: 0 }
        awaiting_sync_count: { type: integer, minimum: 0 }
        sync_failed_count: { type: integer, minimum: 0 }
        synced_count: { type: integer, minimum: 0 }
        shipment_count: { type: integer, minimum: 0 }
        shipped_quantity: { $ref: '#/components/schemas/DecimalQuantity' }
        average_tracking_hours: { type: number, format: double, minimum: 0 }
...